+++ packages/mint-components/CHANGELOG.md
@@ -11,11 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Changed component:
+- Changed components:
   - \<sqm-portal-change-password> Swapped location of cancel and change password buttons, added padding between fields
   - \<sqm-portal-change-password> Changed hardcoded text into customizable props
   - \<sqm-portal-profile-view> Changed hardcoded text into customizable props
   - \<sqm-portal-reset-password> Changed hardcoded text into customizable props
+  - \<sqm-portal-register> Changed hardcoded text into customizable props
 
 ## [1.2.0] - 2021-09-17